requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0

pip~=23.2.1
wheel~=0.41.2
requests~=2.31.0
setuptools~=68.2.0
celery~=5.3.4